swap to using react-select rather than extending widget
